requirements.txt
pytrends==4.9.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.21.0
pandas==2.2.2
pyyaml==6.0.1